add include <QDateTime>

+++ src/Map/OnlineRasterMapLayerProvider_P.cpp
@@ -8,6 +8,7 @@
 #include <QNetworkRequest>
 #include <QNetworkReply>
 #include <QFile>
+#include <QDateTime>
 
 #include "ignore_warnings_on_external_includes.h"
 #include <SkData.h>
@@ -81,8 +82,8 @@ bool OsmAnd::OnlineRasterMapLayerProvider_P::obtainData(
     bool isExists = localFile.exists();
     if (isExists && source->expirationTimeMillis != -1)
     {
-        long currentTime = static_cast<long>(QDateTime::currentMSecsSinceEpoch());
-        long lastModifiedTime = static_cast<long>(localFile.lastModified().toUTC().toMSecsSinceEpoch());
+        long currentTime = QDateTime::currentMSecsSinceEpoch();
+        long lastModifiedTime = localFile.lastModified().toUTC().toMSecsSinceEpoch();
         long expirationTime = source->expirationTimeMillis;
         if (currentTime - lastModifiedTime > expirationTime)
             isExpired = true;
